added time handling functions

+++ libklio/sensor.cpp
@@ -1,3 +1,23 @@
+/**
+ * This file is part of libklio.
+ *
+ * (c) Fraunhofer ITWM - Mathias Dalheimer <dalheimer@itwm.fhg.de>, 2010
+ *
+ * libklio is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * libklio is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with libklio. If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
 #include "sensor.hpp"
 #include <sstream>
 #include <boost/uuid/uuid_io.hpp>
@@ -11,7 +31,7 @@ const std::string Sensor::uuid_string() const {
 const std::string Sensor::str() {
   std::ostringstream oss;
   oss << "Sensor " << _name << "(" << to_string(_uuid) << "), unit " 
-    << _unit << ", tz=" << (int)_timezone;
+    << _unit << ", tz=" << _timezone;
   return oss.str();
 }
 
@@ -27,3 +47,5 @@ bool Sensor::operator == (const Sensor& rhs) {
 bool Sensor::operator != (const Sensor& rhs) {
   return not operator==(rhs);
 }
+
+
